fixnew

+++ evaluator/views.py
@@ -1493,8 +1493,8 @@ class MakeAssignedQuatationPhase1(IsEvaluator,View):
 	def post(self,request,enquiry_id,evaluation_id):
 		
 		payment_method = request.POST.get('payment_method')
-		before_cleaning_amount	= int(request.POST.get('before_cleaning_amount')or 0)
-		after_cleaning_amount	= int(request.POST.get('after_cleaning_amount')or 0)
+		before_cleaning_amount	= float(request.POST.get('before_cleaning_amount')or 0)
+		after_cleaning_amount	= float(request.POST.get('after_cleaning_amount')or 0)
 		
 		#update payment method
 		Evaluation.objects.filter(id=evaluation_id,is_active=True).update(payment_method=payment_method,quatation_status='PENDING',before_cleaning_amount=before_cleaning_amount,after_cleaning_amount=after_cleaning_amount)
